Eric Sunshine eliminated compilation warning reported by Apple/Tiger gcc4.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@21498 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/sprcal3d/object/sprcal3d.cpp
@@ -1744,7 +1744,7 @@ void csSpriteCal3DMeshObject::SetupRenderMeshes ()
 
       csRef<iRenderBufferAccessor> accessor (
 	csPtr<iRenderBufferAccessor> (new BaseAccessor (this, (int)m, (int)s)));
-      rm.buffers->SetAccessor (accessor, CS_BUFFER_ALL_MASK);
+      rm.buffers->SetAccessor (accessor, (uint32)CS_BUFFER_ALL_MASK);
 
       allRenderMeshes.Push (&rm);
     }
